Successfully run advirl with ddpg
Despite the code can run without error, there are several problems with current
implementation:

- For malib/agent/agent_interface.py:L341, I request the data from all
the datasets not be None to fix a bug in adverarial training. However, in
offline training such as behavior cloning, the rollout dataset can be empty.
I think we can add a way to remove the main environment dataset in
offline training.

- The settings update rule now adds the newly specified key after the
default key if the settings item is a dictionary. So we need to manually
set the default value in malib/agent/agent_interface.py:L130.

- The GAIL+DDPG training can not converge.